requirements.txt
Flask~=2.0.1
flask-redis~=0.4.0
flask_sqlalchemy~=2.5.1
google-cloud-secret-manager~=2.7.1
gunicorn~=20.1.0
